requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
gunicorn==21.2.0
